Ats Score Resume using python flask/resume/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
